requirements.txt
paho-mqtt
counterfit-connection
counterfit-shims-grove
counterfit-shims-seeed-python-dht
uuid
